+++ files/default/tests/minitest/default_test.rb
@@ -16,11 +16,11 @@
 
 require File.expand_path('../support/helpers', __FILE__)
 
-describe 'postgresql::default' do
+describe 'rackspace_postgresql::default' do
   include Helpers::Postgresql
 
   it 'installs the postgresql client packages' do
-    node['postgresql']['client']['packages'].each do |pkg|
+    node['rackspace_postgresql']['client']['packages'].each do |pkg|
       package(pkg).must_be_installed
     end
   end