+++ recipes/apt_pgdg_postgresql.rb
@@ -1,18 +1,18 @@
-if not %w(etch lenny lucid precise sid squeeze wheezy).include? node['postgresql']['pgdg']['release_apt_codename']
-  raise "Not supported release by PGDG apt repository"
+unless %w(etch lenny lucid precise sid squeeze wheezy).include? node['rackspace_postgresql']['pgdg']['release_apt_codename']
+  fail 'Not supported release by PGDG apt repository'
 end
 
-include_recipe 'apt'
+include_recipe 'rackspace_apt'
 
-file "remove deprecated Pitti PPA apt repository" do
+file 'remove deprecated Pitti PPA apt repository' do
   action :delete
-  path "/etc/apt/sources.list.d/pitti-postgresql-ppa"
+  path '/etc/apt/sources.list.d/pitti-postgresql-ppa'
 end
 
-apt_repository 'apt.postgresql.org' do
+rackspace_apt_repository 'apt.postgresql.org' do
   uri 'http://apt.postgresql.org/pub/repos/apt'
-  distribution "#{node['postgresql']['pgdg']['release_apt_codename']}-pgdg"
-  components ['main', node['postgresql']['version']]
+  distribution "#{node['rackspace_postgresql']['pgdg']['release_apt_codename']}-pgdg"
+  components ['main', node['rackspace_postgresql']['version']]
   key 'http://apt.postgresql.org/pub/repos/apt/ACCC4CF8.asc'
   action :add
 end