+++ recipes/client.rb
@@ -1,10 +1,12 @@
 #
-# Cookbook Name:: postgresql
+# Cookbook Name:: rackspace_postgresql
 # Recipe:: client
 #
 # Author:: Joshua Timberman (<joshua@opscode.com>)
 # Author:: Lamont Granquist (<lamont@opscode.com>)
+# Author:: Matthew Thode (<matt.thode@rackspace.com>)
 # Copyright 2009-2011 Opscode, Inc.
+# Copyright 2014 Rackspace, US Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,19 +21,19 @@
 # limitations under the License.
 #
 
-if node['postgresql']['version'].to_f > 9.1 && platform_family?('ubuntu', 'debian')
-  node.default['postgresql']['enable_pgdg_apt'] = true
+if node['rackspace_postgresql']['version'].to_f > 9.1 && platform_family?('ubuntu', 'debian')
+  node.default['rackspace_postgresql']['enable_pgdg_apt'] = true
 end
 
-if(node['postgresql']['enable_pgdg_apt'])
-  include_recipe 'postgresql::apt_pgdg_postgresql'
+if node['rackspace_postgresql']['enable_pgdg_apt']
+  include_recipe 'rackspace_postgresql::apt_pgdg_postgresql'
 end
 
-if(node['postgresql']['enable_pgdg_yum'])
-  include_recipe 'postgresql::yum_pgdg_postgresql'
+if node['rackspace_postgresql']['enable_pgdg_yum']
+  include_recipe 'rackspace_postgresql::yum_pgdg_postgresql'
 end
 
-node['postgresql']['client']['packages'].each do |pg_pack|
+node['rackspace_postgresql']['client']['packages'].each do |pg_pack|
 
   package pg_pack