+++ recipes/config_initdb.rb
@@ -1,7 +1,8 @@
 #
-# Cookbook Name:: postgresql
+# Cookbook Name:: rackspace_postgresql
 # Recipe:: config_initdb
 # Author:: David Crane (<davidc@donorschoose.org>)
+# Author:: Matthew Thode (<matt.thode@rackspace.com>)
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -31,7 +32,7 @@
 # original initdb source code.
 #
 # By examining the system configuration, this recipe will set the
-# following node.default['postgresql']['config'] attributes:
+# following node.default['rackspace_postgresql']['config'] attributes:
 #
 # - Locale and Formatting -
 #   * datestyle
@@ -49,12 +50,12 @@
 # logging settings that initdb provided. These settings do differ from
 # the PostgreSQL default settings, which would log to stderr only. The
 # initdb settings rotate 7 days of log files named postgresql-Mon.log,
-# etc. through these node.default['postgresql']['config'] attributes:
+# etc. through these node.default['rackspace_postgresql']['config'] attributes:
 #
 # - Where to Log -
 #   * log_destination = 'stderr'
 #   * log_directory = 'pg_log'
-#   * log_filename = 'postgresql-%a.log'
+#   * log_filename = 'rackspace_postgresql-%a.log'
 #     (Default was: postgresql-%Y-%m-%d_%H%M%S.log)
 #   * logging_collector = true # on
 #     (Turned on to capture stderr logging and redirect into log files)
@@ -69,25 +70,25 @@
 # Locale Configuration
 
 # See libraries/default.rb for the locale_date_order() method.
-node.default['postgresql']['config']['datestyle'] = "iso, #{locale_date_order()}"
+node.default['rackspace_postgresql']['config']['datestyle'] = "iso, #{locale_date_order}"
 
 # According to the locale(1) manpage, the locale settings are determined
 # by environment variables according to the following precedence:
 # LC_ALL > (LC_MESSAGES, LC_MONETARY, LC_NUMERIC, LC_TIME) > LANG.
 
-node.default['postgresql']['config']['lc_messages'] =
-  [ ENV['LC_ALL'], ENV['LC_MESSAGES'], ENV['LANG'] ].compact.first
+node.default['rackspace_postgresql']['config']['lc_messages'] =
+  [ENV['LC_ALL'], ENV['LC_MESSAGES'], ENV['LANG']].compact.first
 
-node.default['postgresql']['config']['lc_monetary'] =
-  [ ENV['LC_ALL'], ENV['LC_MONETARY'], ENV['LANG'] ].compact.first
+node.default['rackspace_postgresql']['config']['lc_monetary'] =
+  [ENV['LC_ALL'], ENV['LC_MONETARY'], ENV['LANG']].compact.first
 
-node.default['postgresql']['config']['lc_numeric'] =
-  [ ENV['LC_ALL'], ENV['LC_NUMERIC'], ENV['LANG'] ].compact.first
+node.default['rackspace_postgresql']['config']['lc_numeric'] =
+  [ENV['LC_ALL'], ENV['LC_NUMERIC'], ENV['LANG']].compact.first
 
-node.default['postgresql']['config']['lc_time'] =
-  [ ENV['LC_ALL'], ENV['LC_TIME'], ENV['LANG'] ].compact.first
+node.default['rackspace_postgresql']['config']['lc_time'] =
+  [ENV['LC_ALL'], ENV['LC_TIME'], ENV['LANG']].compact.first
 
-node.default['postgresql']['config']['default_text_search_config'] =
+node.default['rackspace_postgresql']['config']['default_text_search_config'] =
   case ENV['LANG']
   when /da_.*/
     'pg_catalog.danish'
@@ -130,19 +131,19 @@ node.default['postgresql']['config']['default_text_search_config'] =
 # defaults for the postgresql.cof settings. If the timezone cannot be
 # identified, do as initdb would do: leave it unspecified so PostgreSQL
 # uses it's internal default of GMT.
-tzdirpath = pg_TZDIR() # See libraries/default.rb
+tzdirpath = pg_tzdir # See libraries/default.rb
 default_timezone = select_default_timezone(tzdirpath) # See libraries/default.rb
-if !default_timezone.nil?
-    node.default['postgresql']['config']['log_timezone'] = default_timezone
-    node.default['postgresql']['config']['timezone'] = default_timezone
+unless default_timezone.nil?
+  node.default['rackspace_postgresql']['config']['log_timezone'] = default_timezone
+  node.default['rackspace_postgresql']['config']['timezone'] = default_timezone
 end
 
 #######
 # - Where to Log -
-node.default['postgresql']['config']['log_destination'] = 'stderr'
-node.default['postgresql']['config']['log_directory'] = 'pg_log'
-node.default['postgresql']['config']['log_filename'] = 'postgresql-%a.log'
-node.default['postgresql']['config']['logging_collector'] = true # on
-node.default['postgresql']['config']['log_rotation_age'] = '1d'
-node.default['postgresql']['config']['log_rotation_size'] = 0
-node.default['postgresql']['config']['log_truncate_on_rotation'] = true # on
+node.default['rackspace_postgresql']['config']['log_destination'] = 'stderr'
+node.default['rackspace_postgresql']['config']['log_directory'] = 'pg_log'
+node.default['rackspace_postgresql']['config']['log_filename'] = 'rackspace_postgresql-%a.log'
+node.default['rackspace_postgresql']['config']['logging_collector'] = true # on
+node.default['rackspace_postgresql']['config']['log_rotation_age'] = '1d'
+node.default['rackspace_postgresql']['config']['log_rotation_size'] = 0
+node.default['rackspace_postgresql']['config']['log_truncate_on_rotation'] = true # on