+++ recipes/default.rb
@@ -1,8 +1,9 @@
 #
-# Cookbook Name:: postgresql
+# Cookbook Name:: rackspace_postgresql
 # Recipe:: default
 #
 # Copyright 2009, Opscode, Inc.
+# Copyright 2014, Rackspace, US Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,4 +18,4 @@
 # limitations under the License.
 #
 
-include_recipe "postgresql::client"
+include_recipe 'rackspace_postgresql::client'